+++ src/main/java/com/sparta/myselectshop/aop/Timer.java
@@ -1,0 +1,12 @@
+package com.sparta.myselectshop.aop;
+
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.TYPE, ElementType.METHOD})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Timer {
+}
